Move SnakeEngine to appropriate location and reorder action creators
