+++ penny-guess-client/src/main/Snake/engine/SnakeEngine.js
